# apps/api-gateway/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/buildings": {
      "get": {
        "tags": ["buildings"],
        "summary": "Get paginated list of buildings (public)",
        "operationId": "ListingsController_findAll",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number"
            },
            "description": "Page number"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number"
            },
            "description": "Items per page"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Search query"
          },
          {
            "name": "currency",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Currency filter"
          }
        ],
        "responses": {
          "200": {
            "description": "List of buildings retrieved successfully"
          },
          "400": {
            "description": "Invalid query parameters"
          },
          "503": {
            "description": "Service unavailable"
          }
        }
      }
    },
    "/api/v1/buildings/{id}": {
      "get": {
        "tags": ["buildings"],
        "summary": "Get building by ID (public)",
        "operationId": "ListingsController_findOne",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Building ID (UUID)"
          },
          {
            "name": "currency",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Currency for price conversion"
          }
        ],
        "responses": {
          "200": {
            "description": "Building retrieved successfully"
          },
          "400": {
            "description": "Invalid UUID format"
          },
          "404": {
            "description": "Building not found"
          },
          "503": {
            "description": "Service unavailable"
          }
        }
      }
    },
    "/api/v1/admin/buildings": {
      "get": {
        "tags": ["admin-buildings"],
        "summary": "Get paginated list of buildings (admin)",
        "operationId": "AdminListingsController_findAll",
        "security": [
          {
            "admin-key": []
          }
        ],
        "parameters": [
          {
            "name": "x-admin-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Admin API key"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number"
            },
            "description": "Page number"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number"
            },
            "description": "Items per page"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Search query"
          }
        ],
        "responses": {
          "200": {
            "description": "List of buildings retrieved successfully"
          },
          "400": {
            "description": "Invalid query parameters"
          },
          "401": {
            "description": "Unauthorized - admin key required"
          },
          "503": {
            "description": "Service unavailable"
          }
        }
      },
      "post": {
        "tags": ["admin-buildings"],
        "summary": "Create a new building (admin)",
        "operationId": "AdminListingsController_create",
        "security": [
          {
            "admin-key": []
          }
        ],
        "parameters": [
          {
            "name": "x-admin-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Admin API key"
          }
        ],
        "requestBody": {
          "description": "Building data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Building created successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized - admin key required"
          },
          "503": {
            "description": "Service unavailable"
          }
        }
      }
    },
    "/api/v1/admin/buildings/{id}": {
      "get": {
        "tags": ["admin-buildings"],
        "summary": "Get building by ID (admin)",
        "operationId": "AdminListingsController_findOne",
        "security": [
          {
            "admin-key": []
          }
        ],
        "parameters": [
          {
            "name": "x-admin-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Admin API key"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Building ID (UUID)"
          }
        ],
        "responses": {
          "200": {
            "description": "Building retrieved successfully"
          },
          "400": {
            "description": "Invalid UUID format"
          },
          "401": {
            "description": "Unauthorized - admin key required"
          },
          "404": {
            "description": "Building not found"
          },
          "503": {
            "description": "Service unavailable"
          }
        }
      },
      "put": {
        "tags": ["admin-buildings"],
        "summary": "Update building by ID (admin)",
        "operationId": "AdminListingsController_update",
        "security": [
          {
            "admin-key": []
          }
        ],
        "parameters": [
          {
            "name": "x-admin-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Admin API key"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Building ID (UUID)"
          }
        ],
        "requestBody": {
          "description": "Updated building data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Building updated successfully"
          },
          "400": {
            "description": "Invalid input data or UUID format"
          },
          "401": {
            "description": "Unauthorized - admin key required"
          },
          "404": {
            "description": "Building not found"
          },
          "503": {
            "description": "Service unavailable"
          }
        }
      },
      "delete": {
        "tags": ["admin-buildings"],
        "summary": "Delete building by ID (admin)",
        "operationId": "AdminListingsController_remove",
        "security": [
          {
            "admin-key": []
          }
        ],
        "parameters": [
          {
            "name": "x-admin-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Admin API key"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Building ID (UUID)"
          }
        ],
        "responses": {
          "200": {
            "description": "Building deleted successfully"
          },
          "400": {
            "description": "Invalid UUID format"
          },
          "401": {
            "description": "Unauthorized - admin key required"
          },
          "404": {
            "description": "Building not found"
          },
          "503": {
            "description": "Service unavailable"
          }
        }
      }
    }
  },
  "info": {
    "title": "API Gateway",
    "description": "API Gateway for New Building Portal",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "admin-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-key",
        "description": "Admin API key for admin endpoints"
      }
    },
...
  }
}